build.py: allow CC override for clang
In a similar way that the environment can override CROSS_COMPILE, allow
the caller to override CC.  This enables builds using clang simply by
setting CC=clang when calling build.py, for example:

 ARCH=arm64 CC=clang CROSS_COMPILE=aarch64-linux-gnu- build.py -c defconfig

While at it generalize how the compiler verion is detected to work with
both gcc and clang.

Also, for non-gcc builds, append the CC value to the build output
dir (KBUILD_OUTPUT) and the the ccache directory name.

Signed-off-by: Kevin Hilman <khilman@baylibre.com>